chore: move expresson logic in same folder
