+++ src/parser/expression/ExpressionStep.ts
@@ -1,4 +1,4 @@
-import { Step } from "./step"
+import { Step } from "../step"
 
 export class ExpressionStep {
